Update isLinkHandlingAllowed to mean all or nothing
This toggle is actually used to toggle whether or not the app can open
app links at all, not just verified app links.

Updates the persistence/data classes to be in line with the new default
behavior.

Also makes it so that unverified apps cannot be enabled if another
approved package already exists.

Bug: 178525735

Test: atest DomainVerificationPersistenceTest

Change-Id: I5e258e230e6d6b5de79aab32838496f2126f8451

Former-commit-id: ebf59827eedbc7c6f8add8c32abb8c9bed791054