Controls - Prevent action while locked
After tapping on a control to change the state (lock/unlock, turn
on/off) while the device is locked, the bouncer will be shown. The
action is captured and is supposed to run on the next refresh of the
control, after the phone is unlocked. However, if a well timed refresh
is requested from the controls provider while the user is waiting on the
security bouncer, the action will be fulfilled without the device
being unlocked.

Bug: 179795856
Test: atest ControlActionCoordinatorImplTest
Change-Id: Ie100954a34f19ea6936fe29cbcb205de14f1513c

Former-commit-id: d832feaa41550238c0bd6e3cc5bdd70fafc33719