Remove SurfaceSession.nativeKill
SurfaceSession.nativeKill would call SurfaceComposerClient.dispose. This
was unsafe since it would only clear the ISurfaceComposerClient binder
object, but the SurfaceComposerClient object could still be around.
Instead, when calling SurfaceSession.kill, just remove the Java
reference, but allow the dtor of SurfaceComposerClient to handle
disposing the binder object.

Test: Force app to restart while in a sync transaction. No longer shows
an error
Fixes: 179459565

Change-Id: I322e7f3bd4e7efb3d85d1afe3c4704283f58a985

Former-commit-id: ff265e61777677ce256add37108bef5dae87a42e