BatteryStats.getCustomMeasuredEnergiesMicroJoules
BatteryStats now has access to the custom measured energies,
in the form of an array of energies, indexed by the custom
energy bucket number (i.e. by the ordinal of each energy
consumer of type OTHER).

Also cleans up unused methods (now that we've solidified this
method as being the one required).

Bug: 179107328
Bug: 174818228

Test: atest FrameworksCoreTests:com.android.internal.os.BatteryStatsNoteTest
Test: atest FrameworksCoreTests:com.android.internal.power.MeasuredEnergyStatsTest
Change-Id: I96d646bd844dfa64d48dbb1158358b7561bae971

Former-commit-id: da600e0c4a43ef1e8d3767f11def8ef07f03deb3