Require .dm to install with .fsv_sig if supported
Previously, .fsv_sig for .dm is optional, and is only required if the
base apk has it. This change makes it required regardless of (base) apk,
when a system property is set to "true".

fs-verity is required for devices first launced with R (can also be opted
in). On older devices, .dm is still allowed to install without .fsv_sig.

The change contains 3 conceptual changes:
 - mVerityFound -> mVerityFoundForApks: since it now focuses on only
   APKs.
 - Bail out if .dm does not come with .fsv_sig (on supported devices).
 - Update tests

Bug: 180414192
Test: CtsAppSecurityHostTestCases:android.appsecurity.cts.ApkVerityInstallTest
Test: CtsDexMetadataHostTestCases
Test: ApkVerityTest
Test: Also run the above on a device without fs-verity support, and with
      the requirement disable by system property.

Change-Id: I1c93671b9b433a73b88ae23205eec2a57a147ae1

Former-commit-id: 675088a491e73b6fa9a91ba1512a21ba932f3017