Unsuppress the sensor privacy dialog immediately positive action
This helps in the case that the user needs to disable privacy back to
back for both  camera and microphone. Also this is helpful for testing
so that we don't need to put a sleep between testing both variants of
the dialog.

Also put the logic for delaying the suppression with the UI since this
impacts user experience.

Test: Use GoogleCamera to get it to request mic immediately after
          allowing camera
Bug: 162549680
Change-Id: I63282f1bb5297d5134f1b993d994b5b5a703d624

Former-commit-id: 17762821ed32d72cad2f25e0d2070eeb5d4bfbb1