Fix non-zero window inset
We support 0-width/height window to get shared edge with inset for
b/150696052, but we should limit that to only 0-width/height window.

This will fix the issue that in a dual-display device, when the ime is
only shown on one display, the app on the other display shouldn't get
the inset.

Bug: 175779509
Test: atest InsetsSourceTest
Change-Id: I3c73278a082b084e36be8241c20fa89fb00e405a

Former-commit-id: 418acc10ae4275cb94dfc118be199547e76c9c21