Add authenticator-aware string methods to BiometricManager
Adds methods to BiometricManager to return localized strings that may be
used by apps that utilize biometric authentication. These methods should
provide strings that are appropriate for the specific authenticators
that are present on a device and that will actually be used for
authentication in practice. This should alleviate the issue of apps
needing to use overly broad, less user-friendly strings in order to
cover all possible authentication types that may be used.

Test: atest BiometricManagerTest
Test: Manually checked strings using biometric test app

Bug: 162973174
Bug: 180732913
Change-Id: Icd16d8a3b0339e4a963d47e4eb213fa7df83412c

Former-commit-id: f52c7481925424d47ac348abfe9a4738b3fd52be