Handle the potential recursions within updateOomAdj
Some code paths within the updateOomAdj* would potentially call into
the updateOomAdj* again and result in undefined behavior. This CL
adds the handling for it - in case of recursion is encountered,
the oomAdj update request will be enqueued, and at the end of the
current pass of updateOomAdj*, we'll do another updateOomAdj* to
handle those requests.

Bug: 167773591
Bug: 162476856
Test: atest CtsAppTestCases
Test: atest MockingOomAdjusterTests
Change-Id: I4788ba12f06744204c68382b399216f232059138

Former-commit-id: 64f88fa17344ecea8c9503c9f0ba95a8f9e754bb