Update the order in which we apply new proc_state/oom_adj values.
We check if the network blocked state (which depends on proc_state)
of an uid changed and update a seq counter to keep track of that.
We need to do this before the newly computed proc_state/oom_adj
values are applied. It used to be like this earlier but seems there
was a regression.

Fixes: 181013898
Test: atest ./tests/cts/hostside/src/com/android/cts/net/HostsideRestrictBackgroundNetworkTests.java
Test: atest ./services/tests/servicestests/src/com/android/server/am/ActivityManagerServiceTest.java
Test: atest ./services/tests/servicestests/src/com/android/server/am/OomAdjusterTests.java
Test: atest ./tests/app/src/android/app/cts/ActivityManagerProcessStateTest.java
Change-Id: Ib5b26c82495856a83cbe6d76594e7503ae37f98c

Former-commit-id: 1e7aefbef574a823d7acfcd553b51f4d700b0371