Change deleteRootTasks so that it returns false instead of exception
Before the CL deleteRootTasks throws a RuntimeException when a specified
task has already been GCed. Callers of this API don't have a way of
guaranteeing the existance of task in the system server process, so they
need to catch the excpetions always, which resulted in hiding other
RuntimeException potentially caused by programming errors.

Because missing tasks is not a programming error, this CL changes the
exception to a false return value.

Bug: 179324112
Test: None
Change-Id: I335c43954d467a6bc99c2c958c54a76a48fd2ac9

Former-commit-id: 8770248f7796a78c93f2a5f13cb9831856978f5b