Remove the hidden ConnectivityManager dependency in NMS
NetworkManagementService that is outside the incoming
connectivity module should stop using hidden NETID_UNSET
in ConnectivityManager. The only usage in NMS is
setDnsForwarders() method which is previously used by CS.
The reference is now replaced by calling INetd directly.
The method could be removed now, so remove it to resolve
the dependency.

Bug: 172183305
Test: m ; atest FrameworksNetTests
Change-Id: I63137b240c4938b082c460bb1a25b447e7f29e80

Former-commit-id: 7c7ca85882aaa237ca332097231118d45e50c6a8