Allow filter configurations to have NULL filterSettings
In Tuner java, we mark Settings in all the FilterConfiguration
as nullable. It could even be set null for the filter subtypes
that could have filter settings, for example, the clients could
config TS AV filter with a null settings or a non-null AvSettings.

In tuner JNI, we read settings by default, which could cause crash

Changing the implementation to allow the filter configurations to
have null settings

Test: atest android.media.tv.tuner.cts
Bug: 181152149
Change-Id: I7c1d8721aae6da8e557138537e65966628584a9a

Former-commit-id: ea05b4b86f6ae62848b31b09fea7c6319c5a30ee