Launch activity intents on the StatusBar's display
Add the activityOptions setCallerDisplayId and setLaunchDisplayId to
the display the StatusBar is currently on, so that the resulting
activity will be started in the same display as the user is interacting
with.

Bug: 175329765
Test: Manually tested using Exo. Tapping the notification on the Exo
brings it to the Exo virtual display, and tapping on the phone's native
status bar brings the activity back to the primary display.

Change-Id: I2f950131fa4d3aa1a9b423ae65902be68a628178

Former-commit-id: 8d1c1abbcd5e9646ee93f628b75c759c9e10eb4c