General BytesMatcher utility class.
As part of an upcoming feature, we're anticipating the need to match
arbitrary byte sequences, such as Bluetooth UUIDs or MAC addresses,
so this change adds a new BytesMatcher similar in spirit to
the existing PatternMatcher.

It supports a reasonably compact serialization format which can be
used to transport a server-pushed deny-list.  The choice to support
both "accept" and "reject" rules in an ordered list means that very
complex matching rules can be expressed efficiently.

Bug: 181812624
Test: atest FrameworksCoreTests:android.os.BytesMatcherTest
Change-Id: I55193051354278abaef8bef51413c632b07c7177

Former-commit-id: 6644436a75353ace2560cc8d72b01f4ebb0480f6