Fix unbounded MotionEvent growth in Falsing
We were keeping an unbounded list of MotionEvents inside of the
FalsingManager. Generally, this list should have shrunk itself
periodically, but in practice it only shrank itself when someone
double taps a notification.

Meanwhile, we weren't even using the data in the list!

With this change, we only keep the current and prior list of
MotionEvents - the ones we actually use. The list of historical
events is removed, so the leak goes with it.

Fixes: 177329773
Test: atest SystemUITests && manual
Change-Id: If8bfeea944850d972434915018b5bbfb3acb80e4

Former-commit-id: 7da60de0908b0024c076d27b5661d1094b4526ee