getDmabufMappedSizeKb: Use ReadDmaBufMapRefs()
To allow clients more flexibility AppendDmaBufInfo() is being removed.
It allowed reading both mapped and FD-referenced dmabufs, or just mapped
dmabufs if read_fdrefs was specified as false. Internally the function
called ReadDmaBufMapRefs() and ReadDmaBufFdRefs().

This change replaces AppendDmaBufInfo(read_fdrefs=false) with a call
to the underlying ReadDmaBufMapRefs() api which is now made public.

Bug: 181976677
Test: dmabufinfo_test
Change-Id: I9c05f0be9fc4d9cd40cdb146f3f6ca4860f7cd0f

Former-commit-id: 66d2190ed6bdc6533f48ba4256b200555b9b3031