No longer call registerNativeAllocation
The original rationale is obsolete, since contexts are now almost
always shared.

This call was unbalanced; there was no corresponding
freeNativeAllocation call. Thus, in the rare cases in which this
actually makes a difference, it's likely to lead to perpetually
increasing native allocation counts, which is more likely to be
confusing than helpful to the garbage collector.

(Discussed with Tim and Jean-Luc, and concluded that it was best
just to remove the code.)

Test: Treehugger
Bug: 181351667
Change-Id: I376a16732231aad0fdacd022de301464e5ac71dd

Former-commit-id: 5dc6b62136e03fd34ca6979369351198d23531b4