Define new NEARBY_DEVICES permission group
This defines a new Android permission group named NEARBY_DEVICES,
which is displayed to the user as "Nearby Bluetooth Devices".

This also defines two new permissions:
- BLUETOOTH_SCAN
- BLUETOOTH_CONNECT

The BLUETOOTH_SCAN permission is a new dangerous permission which will
henceforth be used by applications that wish to be allowed to scan for
or pair with Bluetooth devices.

The BLUETOOTH_CONNECT permission, likewise, is used to connect to and
communicate with already-paired devices.

(As of this commit, these permissions are not yet used, and the
corresponding AppOps are not yet mapped, so these changes do not yet
have any effect.)

Bug: 181813006
Test: manual
Change-Id: I05c82ae05813189a5ec41656a15c6f4fc770ca57

Former-commit-id: 11ab8b2042acb1fa126fef2a81d7674538b59e44