Add CollapsingToolbarBaseFragment for implementing collapsing toolbar.
In apps following the pattern to do all the things in fragments, and
possibly loading different fragments for different form factors, their
activity would need to remain a stub and can't contain phone-only
layout, making CollapsingToolbarBaseActivity unsuitable for them. So
add a CollapsingToolbarBaseFragment for integrating the same
functionality in fragments.

The layout files are also moved to be under the same name, but
different content depending on the platform version, to simplify the
logic loading them.

Bug: 177638625
Test: manual
Change-Id: Ie9e2a147dd4bd2c1ee3924b1afc4c10a6f735e59

Former-commit-id: e1f6cb431e3f37d33c35992a9cdbee8f57f6d5c6