Per-window input rotation: Remove projection from DisplayViewport
InputFlinger always needs to know about the display orientation because
some devices (such as mice and touchpads) are orientation-dependent.

Instead of removing the orientation for DisplayViewport, we remove the
display projection so that the logical, physical, and device frames are
the same. The display projection will be included in the per-window
transform set by SurfaceFlinger.

Bug: 179274888
Test: test touch, mouse, and touchpad input in all four orientations
using a test app and the PointerLocationView.

Change-Id: I00481c67ee6dfa93115b299999ea7c0d4baa8e02

Former-commit-id: 3644df517456103807b56cf8022a4ca054e41db6