Rely on a falsing-belief event to reset lockscreen
Individual components of the lock screen are now simply in charge of
asking the FalsingManager to recognize certain types of gestures. If
they don't ask, the FalsingManager will assume they were accidental.

Once a certain number of accidental touches occur, an event is fired
that signals that the lock screen should be reset.

More tuning and adjustment to come, but this is now functional.

This also fixes a bug in an earlier cl where the bouncer would no
longer retract if touches outside of its region were detected.

Bug: 172655679
Test: manual
Change-Id: Iffd1c7e58365235b1b7dbdac0c74439ead1df9d3

Former-commit-id: 39ed7729af6016a1c7b3174951ab21186933dfa0