Remove pin_split_apks flag
Now the toolchain side fix is in (b/174697187) and it's ready to
pin split APKs, so this doesn't have to be configurable.
(It was behind the flag in the case the other side of the fix
 doesn't happen before the OS release.)

Note that only pinnable app that with split APK so far
is Google Camera.

Bug: 174698005
Test: * Install GoogleCamera over PlayStore, then:
      * Run following command and inspect the output:
 $ adb shell device_config put runtime_native_boot pin_camera true
 $ adb shell cmd pinner repin && sleep 1 && adb shell dumpsys pinner
Change-Id: I20b4cf99aff6d55b813b48ed921bb9c25c06ae51

Former-commit-id: 637c52b861d36f022350d252c5c214dbb07e81b9