Remove network transport from core/proto
Move the network transport type to proto_logging. It is needed in
atoms.proto, and all atoms.proto dependencies should live in
proto_logging. Previously, networkcapabilities.proto was in
proto_logging, but the majority of that proto is not needed for logging
and is used for incident reports.

Test: m
Bug: 182392879
Change-Id: I24ff96806e417ad4ee89d84f051ea790a416338f
Merged-In: I24ff96806e417ad4ee89d84f051ea790a416338f

Former-commit-id: 80d273e43da6b3c1281a72f673da1322adf9cc6e