Use BLAST layer for screen rotation, surface freezer, and organizer
Use a BLAST layer when taking a screenshot and using it for an
animation. This is simple enough to do with BLAST instead of a buffer
queue since we can just submit a single buffer and it will be presented.

Added new methods in SurfaceControl to support sending a buffer and
colorSpace.

Test: ScreenRotation
Test: Freeform animation
Bug: 170765639
Change-Id: I72081ab833d76fac40667eed52cf16798112a5b6

Former-commit-id: 0bfc89b9fedf376fb6a761e984b26d38fc7118ee