Don't skip IME animation when backed to Task has no editor focus
When the dialer app previously shows IME, if the user swipe out the
dialer task to launcher or switching task and then navigating back
to dialer task, will see IME surface will fade-out because the
editor view has been cleared the focus.

The issue happens when user taping editText to show keyboard and
the IME show animaion will be skipped.

It could be related the CL[1] that we skip the animation once
when the task snapshot has the IME surface, but we should not
skip IME show animation if the launching task has been cleared
the editor focus.

[1]: I83ffc03119e01da71ad12f8ad8043cf0730dfd50

Fix: 183347520
Bug: 166736352
Test: manual as steps in commit message
Test: atest ImeInsetsSourceConsumerTest#\
      testImeGetAndClearSkipAnimationOnce_expectNoSkip
Change-Id: Ib140801f1ce03b5566e756914f96dacba3ad8892

Former-commit-id: 5a810c8403eeafc6e86084c6b0ce11c96908a30d