Fix and move old One Touch Play tests
Moves the One Touch Play tests in HdmiControlServiceBinderApiTest to
OneTouchPlayActionTest. This fixes tests failures on ADT-3, which were
caused by the test setup responding with the <Report Power Status>
message on same callstack as the <Give Device Power Status> message,
causing it to be processed before OneTouchPlayAction updated its state.

Bug: 182910754
Test: atest com.android.server.hdmi.OneTouchPlayActionTest
Change-Id: I11e8e2a5259739325d1dd9aae92113e85ecad47a

Former-commit-id: 807dfb306788593f7ebd6f07c268c7b6310275ba