Detects all activities for whether showing work challenge
Work challenge did not show when a work activity is not on top, but
still visible after screen turns on.

Also show work challenge even if the work activity is behind a top
fullscreen activity of another profile because the user can still
navigate back to the work activity when top activity finishes.

Bug: 168527567
Test: manually test work challenges
Test: RootWindowContainerTests

Change-Id: I5e09b09be547d04fdfd709cb9cd4bcd4a94bbf21

Former-commit-id: c83234a0ef4180343af8cc3cb0edcdf92eee2a46