Fix inconsistency in ringtone list
Sometimes when ringtone from external storage gets same _id as any ringtone of internal volume, getRingtonePosition(..) matches with the internal one firstly, and the ringtone shown is wrong.

Solution is to  match the URI of ringtone instead of the _ID so that correct ringtone is chosen.

Bug: 172878342
Change-Id: I8fc3e936fe89308d4697eda42088e4fd533ce47c
Merged-In: I8fc3e936fe89308d4697eda42088e4fd533ce47c

Former-commit-id: d9b89825b71407457f51786b2ba5b2a37341416a