Add a tool to dump git SHAs of API changes
Necessary for use by API council near the end of a release to verify
their server-side tool has filed bugs for all the API changes.

This utilized bazel-query to figure out what contributes to current.txt,
and then finds api subdirectories in the relevant directories. Finally,
it prints the non-merge SHAs of changes in those directories.

Test: frameworks/base/api/dump_api_shas.sh
Change-Id: I330bfe24fb4df5a8a9218241540f7a7ab3b66e2f

Former-commit-id: 8caea0d426c4599afb19d1a913252cd5d5176536