Replace BLUETOOTH and BLUETOOTH_ADMIN permission checks in BluetoothManagerService.
This replaces the checks for the legacy bluetooth permissions
with checks for the new BLUETOOTH_CONNECT permission.

All of the new checks are done as ForPreflight in this CL, i.e. app
ops will be checked but not noted. Changes to make some of the
checks ForDataDelivery (i.e. noting the app op) will be made in
follow-ups - this requires additional work as the calling package
and attribution tag must be plumbed through from the app side.

Bug: 183203469
Test: builds
Change-Id: I7b99a203ac25a8dcc0d33f442b08203244bff534

Former-commit-id: 2c8f0f027bdb69056d80952aa8d8c3ef14650ea3