AOD: selectively register sensors that use prox
Allow some devices to only register sensors that use prox when
the display state is off or in doze on certain devices to
avoid white prox dot behind display.

The delay shouldn't be perceptible - when the device isn't in low-power
mode yet, we rely on regular touch from the display.

Test: manual, atest DozeSensorsTest, atest DozeTriggersTests
Fixes: 181835575
Change-Id: Ib62ed0ca3150e0a3939c8d9f0c5ae700fd6a61d7

Former-commit-id: 78386ab1ca78e317278136d940091dc1328fb2c0