Custom haptic composition for assistant launched via button.
The effect is a small ramp, a 50ms gap, and a sharp tick,
evoking the crisp snap of a sail, or a molecule's bonds
approaching and then snapping together.

It might be nice to start this effect slightly early, but
that will require more tweaking to the keypress messaging
and timers. We'll see if it's necessary.

For devices without haptic composer support, fall back to
EFFECT_DOUBLE_CLICK.

Bug: 182381083
Test: manual (literally, you have to hold the device):
      adb shell settings put global power_button_long_press 5
      adb shell input keyevent --longpress POWER
      (or just hold the power button, since you're holding
      the device)
Change-Id: I0633efd46296ca3f34e5df60be37871fb1a9fb5d

Former-commit-id: 8ff8bd327c8e857001ebc53e701b7476111c971a