Set mIsAnimationPending when the leash is not ready
The server might return a control without a leash to a client while the
setup-transaction of the leash has not been applied yet. If the client
tries to play the animation with the control, the animation won't take
effect.

This CL sets mIsAnimationPending while changing the requested visibility
while the leash of the real control (not the fake one in transient mode)
is not ready. So when the client receives the leash later, the animation
will be played.

Fix: 183362710
Test: steps in the bug
Change-Id: I53108bacf98ac76c3f1e46cdae0a98581bef77f5

Former-commit-id: c2568d6c5898f46e04c97c62c75a9c9e713f6940