String for: Failed guest creation doesn't crash Settings
In the new multiuser settings, there's no logic to handle
what happens when a guest user creation fails, and so Settings
crashed with a NullPointerException.
In the fix, we catch this case to avoid crashing, and display
a toast to the user so that they understand the operation was
not successful.

In this cl we create the string for the toast.

Bug: 175042193
Test: manually failed to create user and observe toast and no crash
Change-Id: I5d5e0a26d52ea81db273c2b74e259ad6783b5c97

Former-commit-id: bec782c911c82b6937c2645fedc5229a96418e0a