Ensure that FalsingManager#isFalseTap handles delayed clicks.
Views, be default, don't immediately process touch events as clicks.
Instead, they post an internal runnable to the main thread that then
runs the click. This was tripping up the FalsingManager, as a recent
change meant that it no longer saw the motion events.

With this change, the FalsingManager will look at prior events if
the recent events are empty.

Fixes: 184635871
Test: manual
Change-Id: I69981b09837d8a1d94b3a033243bce7a2d31844c

Former-commit-id: 7a08f2b716dbe310bc63bae3c2fb8212b8a7717f