Adjust geotz feature config logic
A set of related changes under the heading of "configuration logic
changes".

The ability to 'discover' the package of a LTZP has been removed. A
provider's package name must now match the package name held in
config.xml. This takes advantage of changes in ServiceWatcher (which,
due to a recent change, no longer insists on specific configuration
options). It will allow the package name to be determined at runtime via
server flags in a future change, which will make future testing easier.

The overall location time zone detection feature can also now be
configured on and will not "show up" if there are no providers currently
enabled. The ability to have the location time zone detection feature
"supported" on the device be on and not have it "show up" in the UI when
there are no providers will make testing easier when the secondary
provider has been removed from AOSP (bug 178169250).

To remove command line duplication of things that can now be achieved
via device_config (AKA server flags), there have been various command
line changes. It is also now possible to override a provider's "mode"
via the command line to be "enabled". Previously, a provider could only
be disabled or set into a simulation mode via the command line, which
assumed the default state was "enabled" (bug 178169250).

Bug: 175718363
Bug: 178169250
Test: See the associated change in cts/
Change-Id: I0d3314155cef4f122cae0d569b4687b658e777b1

Former-commit-id: 59ff2dca0ee529c00ef0c0892430d9f33296c83c