Remove annotations from generated class until tests updated
Codegen is a tool that generates parcelable classes from
field declarations. Hence, the @SystemApi annotations are
placed on the field and the tool propagetes them to the
generated APIs. We need to update the AnnotationTest to
either ingore these generated classes or ignore private
fields annotated with @SystemApi as they are not API in
practice. This CL removes the annotations to fix the test.

bug: 184086213

Test: atest android.signature.cts.api.AnnotationTest

Change-Id: Ibe9059f20d95d4cda4eb402704e3c4c002ff0e47

Former-commit-id: 79db6e580790e49b44bc826f6b02ee1f6a41d0ca