Create a snapshot cache utility
Bug: 181964615

The SnapshotCache class can be used with a Watchable; it allows a
snapshot to be reused until the Watchable changes.  Caching snapshots
significantly reduces the time required to create a PackageManager
snapshot.

Set ENABLE to false to disable caching if a problem is found.

Two existing cached snapshots use the new class.  The change to
Settings and InstantAppRegistry do not change functionality.

Test: atest
 * FrameworksServicesTests:WatcherTest
 * FrameworksServicesTests:PackageManagerSettingsTests
 * android.appsecurity.cts.InstantAppUserTest

Change-Id: I76d9c95205ef2a358b6493f02ff3f1f8855a2157

Former-commit-id: a5f59b2a9a7f1e55f3060fc4eeb1f24405581c0b