Use default font for clock font
Note: currently the default is roboto and provides limited support
for variable fonts (ie: Latin numerals supported but other locales may
not be supported by default). This means in certain locales, the font
weight may not change between AOD and the lock screen. OEMs should be
aware that this may cause burn-in on AOD.

Test: manual
Fixes: 171376810
Change-Id: Icb61fec9636ebacec5b7c39492423a9d4bca07c1

Former-commit-id: 68d6e8ebb8fc30320d9280c844d3125f14fab7ee