Remove hidden connectivity method access in FrameworksVcnTests
Due to connectivity modularization work, hidden connectivity
methods and members are not accessible outside the module.
Remove the corresponding usage.

The test case in VcnNetworkProviderTest are not able to create
NetworkRequests with request id assigned. The loop to create
different request is removed. After that, the test does not
provide more test coverage than other tests. Thus, Remove the
test case directly.

Bug: 182859030
Test: atest FrameworksVcnTests
Change-Id: I488f62089d1dbe93c232f892885d944bef896df6

Former-commit-id: 18f3a26efad77c1da290c01e67f38b88ca4156d8