Use the existing wallpaper command API to send wake/sleep events to the wallpaper.
This is already used for similar events such as icon drops and user taps. Using it to indicate that the user has woken/slept the device (and passing along location information) makes sense!

This does require an API modification to add the new string constants to WallpaperManager. This change is being requested by 1P apps (the Live Wallpapers, which are not bundled with the OS and thus can't use system APIs).

Bug: 169693662
Test: locally with modified wallpaper, it receives events
Change-Id: I0cc1bc70dbea509a049b08f8119ed875ffc2e8d5

Former-commit-id: bde980759e036c53be6c2e40619fa13bf9ac052e