Increase gesture completion delay for falsing.
The FalsingManager was hitting a race condition with View click
listeners, where clicks would still, somtimes come in after the
FalsingManager had already marked the gesture as un-analyzed.
Increasing the delay from 50ms to 100ms should all but prevent that
from happening.

I have also decreased the penalty for unalayzed taps since
accidentally tapping on the wrong part of the screen is not unlikely.

Fixes: 185212498
Test: manual
Change-Id: I91e1e9152536a36eda74b5fd60051acc78434761

Former-commit-id: 9a984a26a04e1cd49296e4f6bc1b8a117d6deb2c