Create a new config for attentional haptics OGG resource
Create a new hidden config with the URI of a OGG file that has only the
haptics channels and can be played for attentional haptics when sound is
muted (ringtone, notifications, alarms).

First usage introduced is to the Telecom project to be played on
RINGER_MODE_VIBRATE.

Also update Ringtone class to allow playback of haptic-only ringtones
(i.e. ringtones with unmutted haptic channels and volume set to zero on
devices with audio-coupled haptics support).

Bug: 182157749
Test: manual
Change-Id: Ic64d8d56c6b029e32b3d55389c416cee7679ddc8

Former-commit-id: 852a203646a2c996ddcff069f4332a44fb040b30