Fix magnification bounds cut out with floating a11y button
The window type Floating Accessibility Button is
TYPE_NAVIGATION_BAR_PANEL which is unmagnifed, it leads to
the cut out while drawing the magnification bounds.

To fix it we did a compromise. We omit the window bounds of
TYPE_NAVIGATION_BAR_PANEL windows while computing the
magnification bounds. The cons is that magnification region
is really changed. That means the interaction with such
windows might have latency depending on if they are
inside the magnification bounds. However, it is acceptable because
users don't have complex interaction witch such windows and
see them rarely.

bug: 185401123
Test: manually test
Change-Id: Ia0088a0140fb8d5516e703ebfa5e203f857636d9

Former-commit-id: 987144aa227aba76e1806ac6b5a1555f26b18045