Call OnPrimaryClipChanged when classification status changes.
To allow apps to know when text classification has been completed
on clipboard clips, ClipboardService will now call any
OnPrimaryClipChangedListeners when the classification status
of the primary clip changes.

This change is made as per API council suggestion.

This CL also ensures that the classification status is set on
any related profiles.

Bug: 185177537
Test: atest ClipboardManagerListenerTest
Test: atest ClipDescriptionTest
Test: atest atest ManagedProfileCrossProfileTest#testCrossProfileCopyPaste

Change-Id: I63c44a051d1e8029b6d56b9af5d1e506355a8466

Former-commit-id: 7461b08f9fe3dea74bb8b713d0ff94f5ee25fdf2