Fix ActivityOption#setTaskOverlay can be covered by other activity.
Although ActivityRecord#setTaskOverlay has set setAlwaysOnTop, but it
is not useful in Task#getAdjustedChildPosition, because when sorting
activities in a task, it is meanless to check the windowing mode for
an activity. A simple solution is to override the isAlwaysOnTop method.

And because Task#showToCurrentUser will depends on current top task,
so we cannot add the task before adjusting the position.

Bug: 181721787
Test: atest RootTaskTests StartActivityAsUserTests StartActivityTests
Change-Id: I7fdf19d04bf71fa97a841223c06fad9a7e0b2ee4

Former-commit-id: dec663653f3dc107f500fb933fdc7d2b3f5c0676