Handle media session destroyed
It is possible for a media session to be destroyed without updating the
playback state or removing the notification, in which case the
controller will no longer be valid. So we can treat this like
resumption, and re-register the controller and re-activate the controls
if we later get an event where the media is playing.

Fixes: 185899440
Test: MediaTimeoutListenerTest
Change-Id: If8b8922b436261ccf703ee9e1e3c0294171a6c2e

Former-commit-id: 9985b65f641127158e3646588a9f8c58a9dbcb3d