Fix IME switch dialog crash on Dual rootDA display
The crash is because when the IME switch dialog created window token
is removed, it should switch back to listen to ImeContainer.
However, previous CL use the wrong API findAreaForToken, which only
finds the DisplayArea.Token under mAreaForLayer list.
In this scenario, we cannot find one because ImeContainer is moved to
its sub-RootDisplayArea.

This CL updates the naming and add more javaDoc to prevent from
confusion.

Test: atest WindowContextListenerControllerTests
Test: manual - repro steps in b/186366728#comment1
fixes: 186366728

Change-Id: I50d3247b0a158e83d09be893d31802d673a60697

Former-commit-id: 71cb758286938a27d85be09145859a909144eb23