Add sampling to FGS atom
Only log the FGS atom for a certain set of packages. Each device
will have a random but stable set of apps, determined by the ANDROID_ID.

By default, the sampling rate is 1, so we log all packages.

If the rate is 0.5, we'll long only the half of the packages. Again,
each device will have a random set of the packages, so the same package
will be logged only on half the devices.

Fix: 184867830
Test: atest FrameworksServicesTests:ActivityManagerUtilsTest
Change-Id: I4cac13ec62d9c98b95df57d47d63e725ba1fcfdd

Former-commit-id: 2c7fdfcfdac38cb4c0b74ffe22921d2bcaee47bd