Adds the continuous unlock animation!
This adds the new unlock animation, which is a subtle scale and translate from the bottom part of the screen. More importantly, it connects the surface behind the keyguard to your finger when swiping, so that it's one continuous motion between the lock screen and the app/launcher.

Issei's RemoteAnimation-enabling CL was rolled back for a test breakage, so you'll need to enable it to see this:
adb shell setprop persist.wm.enable_remote_keyguard_animation 1

A few P2 issues to be fixed in follow-up CLs since this one is getting large:
- Status bar icons need to animate mid-swipe so they don't jump cut post-unlock (b/183063033)
- Launcher should use a custom animation rather than surface-based animation (b/183063432)
- The scrim should probably animate out once the surface appears rather than at the end (b/183062235)
- Handle RemoteAnimation timeout, either by restarting it, or cancelling the gesture (b/183066204)

Test: change security to swipe, unlock via a slow swipe and a fast swipe and everything in between (also, release swipe halfway)
Test: PIN/password security
Test: double tap to open a notification
Test: fingerprint bypass
Test: fingerprint from lockscreen
Bug: 169692441
Change-Id: Ic78c603b2375d36cf2170b81cca7cddbf334408b

Former-commit-id: ea36ab6533dba657e5395f49deecc0669f79933f